Rename to impact_function_v4
